pyproject.toml
[project]
name = "legalscraper"
version = "0.1.0"
description = "Legal RAG stack (FastAPI + pgvector + Celery worker + data pipeline)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "openai==1.51.2",
    "httpx==0.27.2",
    "psycopg==3.2.1",
    "psycopg-binary==3.2.1",
    "psycopg-pool==3.2.1",
    "pgvector==0.2.5",
    "redis==5.1.1",
    "celery==5.4.0",
    "python-multipart==0.0.12",
    "pypdf==5.3.0",
    "pypdf2",
    "pdfplumber==0.11.5",
    "requests",
    "beautifulsoup4",
    "torch",
    "sentence-transformers",
    "tiktoken",
    "langchain",
    "langchain-openai",
    "pytest",
    "passlib[bcrypt]",
    "python-jose",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# This repo is a monorepo; we don't publish a Python package from the root.
packages = []
